requirements.txt
chex
distrax
dm-haiku
flax
gymnasium
gymnax
ipykernel
ipython
jax
joblib
matplotlib
minatar
numpy<2.0
omegaconf
optax
pandas
pgx==2.4.0
plotnine
pyarrow
pydantic
qdax
ribs[visualize]
scipy
#torch
tqdm
wandb
qdax[cuda12]
MinAtar==1.0.15
gymnasium==0.29.1